Fixes to StmBench7:
* DeuceSTMOperationExecutor => split the atomic block in two separated blocks,
one for the execution of read-only operations and other for read-write. Some
STMs, such as the LSA, keep track of the functions  that perform write
operations and for those ones it will not try to execute read-only transactions.
So to take advantage of the lightweight behavior of read-only transactions we
need to keep a separated method reserved for the execution of read-only
operations, otherwise the STM will not perform read-only transactions.

* BenchThread => this class keeps two multi-dimensional arrays to register the
operations TTC and highest TTC log. Yet, these arrays are thread-local and do not
need to be transactified by Deuce, with STM barriers to access them. So, we moved
them into an auxiliary class ThreadStats and exclude this class from Deuce
instrumentation.

* BenchThread => the system property uniqueop can specify a single operation to
be executed by this benchmark.