JVSTM-lock-free (jvstm-2.0) integration in Deuce
* build.benchmarks.xml => include scripts for the configuration and
execution of the benchmarks with the JVSTM lock free and performing
capture analysis.

* To support Deuce transactional arrays with JVSTM we use the VBox
Array approach, meaning that arrays are wrapped inside a VBox.
However, when we access arrays from a non-transactional scope,
we are not using STM barriers and we cannot deal directly with VBoxes.
So, in this case we unwrap the encapsulated array from the VBox.
Yet, this solution has a limitation regarding multi-dimensional arrays,
due to the big overhead of unwrapping the encapsulated array from
the VBox.
So every time an unwrap operation occurs for a multidimensional array,
then it raises an exception alerting the end-user programmer. In this case,
one of the recommended workarounds is to access those arrays through
atomic indexers (which are performed within a transaction) and avoid
the unwrap. For this purpose we make some minor changes to some
benchmarks replacing the arrays access by atomic indexers.